FIX :: style - ADD :: 팀페이지 백그라운드
