+++ src/components/common/Header/index.tsx
@@ -1,0 +1,31 @@
+import React from 'react';
+import { color } from 'styles/theme';
+import header from 'datas/header.json';
+import Button from 'components/common/Button';
+import Logo from 'assets/logo.svg';
+import * as S from './style';
+
+export default function Header() {
+  return (
+    <S.Header>
+      <S.InnerFrame>
+        <a href="/#main">
+          <S.Logo src={Logo} />
+        </a>
+        <S.Nav>
+          {header.map((item) => {
+            return (
+              <a href={item.to} key={item.id}>
+                <Button
+                  title={item.title}
+                  color={color.gray04}
+                  onClick={() => console.log(item)}
+                />
+              </a>
+            );
+          })}
+        </S.Nav>
+      </S.InnerFrame>
+    </S.Header>
+  );
+}
