+++ src/components/common/RadioGroup/style.ts
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export const RadioGroup = styled.div`
+  display: flex;
+  padding-top: 15px;
+`;
