--- src/components/ui/RadioGroup/index.tsx
@@ -1,10 +1,0 @@
-import { ReactNode } from 'react';
-import * as S from './style';
-
-interface PropsType {
-  children: ReactNode;
-}
-
-export default function RadioGroup({ children }: PropsType) {
-  return <S.RadioGroup>{children}</S.RadioGroup>;
-}
