--- src/components/ui/RadioGroup/style.ts
@@ -1,6 +1,0 @@
-import styled from 'styled-components';
-
-export const RadioGroup = styled.div`
-  display: flex;
-  padding-top: 15px;
-`;
