+++ src/components/ui/Teammate/index.tsx
@@ -1,8 +1,8 @@
 import { useRecoilValue } from 'recoil';
 import { radioState } from 'atoms';
 import TitleWrap from 'components/common/TitleWrap';
-import RadioGroup from 'components/ui/RadioGroup';
-import GenerationRadio from 'components/ui/GenerationRadio';
+import RadioGroup from 'components/common/RadioGroup';
+import GenerationRadio from 'components/common/GenerationRadio';
 import Profile from 'components/common/Profile';
 import generation from 'datas/generation.json';
 import profiles from 'datas/profiles.json';
